Offer choice of audio file format for download
Debian wheezy's avconv does not support audio conversion from webm
format. It crashes with a "...webm: could not find codec parameters"
error.

To solve this issue, we offer the user to specify the preferred audio
file format. If available, this format will be downloaded, instead of
the best possible audio format. Note that in such cases, the audio file
quality may be degraded.

Should close issue #21 (and maybe close issue #22 as well?)